refactor(api): move start-or-resume-competence-evaluation.js
