+++ api/src/evaluation/domain/usecases/start-or-resume-competence-evaluation.js
@@ -1,6 +1,6 @@
-import { Assessment } from '../../../src/shared/domain/models/Assessment.js';
-import { CompetenceEvaluation } from '../models/CompetenceEvaluation.js';
-import { NotFoundError } from '../../domain/errors.js';
+import { Assessment } from '../../../shared/domain/models/Assessment.js';
+import { CompetenceEvaluation } from '../../../../lib/domain/models/CompetenceEvaluation.js';
+import { NotFoundError } from '../../../shared/domain/errors.js';
 
 const startOrResumeCompetenceEvaluation = async function ({
   competenceId,
